+++ src/componenents/Medicine/AllMedicine.js
@@ -2,6 +2,8 @@ import React, { useContext, useEffect, useState } from 'react';
 import { Link, useLoaderData } from 'react-router-dom';
 import SearchBar from '../Allicu/SearchBar';
 import { AuthContext } from '../../Providers/AuthProvider';
+import '@fortawesome/fontawesome-free/css/all.min.css';
+import Swal from 'sweetalert2';
 
 const AllMedicine = () => {
     const medicines = useLoaderData();
@@ -35,18 +37,35 @@ const AllMedicine = () => {
     };
 
     const handleDelete = _id => {
-        console.log('Please delete', _id)
-        fetch(`https://icubd-server.vercel.app/medicine/${_id}`, {
-            method: 'DELETE'
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                console.log('delete', _id);
+                fetch(`http://localhost:5000/medicine/${_id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            Swal.fire(
+                                'Deleted!',
+                                'ICU deleted.',
+                                'success'
+                            )
+                            const remaining = allMedicines.filter(medicine => medicine._id !== _id);
+                            setMedicines(remaining)
+                        }
+                    })
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    alert('Deleted Success', data.name)
-                }
-            })
     }
-
 
     return (
         <div>
@@ -83,10 +102,10 @@ const AllMedicine = () => {
                                         <td className='text-center' style={{ marginBottom: '1rem' }}>{medicine.price}</td>
                                         {
                                             isAdmin && <td>
-                                                <button onClick={() => handleDelete(medicine._id)} className='bg-white p-1 mx-1'>X</button>
+                                                <button onClick={() => handleDelete(medicine._id)}> <i className="fas fa-trash-alt"></i></button>
 
-                                                <Link>
-                                                    <button className='bg-white p-1'>Update</button>
+                                                <Link className='mx-2' to='/'>
+                                                    <i className="fas fa-edit"></i>
                                                 </Link>
                                             </td>
                                         }
